requirements.txt
requests
beautifulsoup4
google-generativeai
pyyaml
python-dotenv
playwright
